refactor store files
