+++ src/store/pin.rs
@@ -1,0 +1,111 @@
+use anyhow::{Context, Result};
+use serde::{Deserialize, Serialize};
+use std::fs;
+use std::io::Write;
+use std::path::Path;
+use std::path::PathBuf;
+use twox_hash::xxhash3_128;
+
+#[derive(Debug, Default, Serialize, Deserialize)]
+pub struct Tree {
+    pub entries: Vec<Entry>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub enum EntryKind {
+    Blob,
+    Tree,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Entry {
+    pub kind: EntryKind,
+    pub hash: String,
+    pub name: String,
+}
+
+pub(crate) fn pin_file(store_path: &Path, file_path: &Path) -> Result<String> {
+    let contents = fs::read_to_string(file_path)?;
+
+    pin_contents(store_path, contents)
+}
+
+pub(crate) fn pin_contents(store_path: &Path, contents: String) -> Result<String> {
+    let hash = xxhash3_128::Hasher::oneshot(contents.as_bytes());
+    let hash = format!("{:032x}", hash);
+    let hash_folder = PathBuf::from(&hash[..2]);
+    let dir = store_path.join(hash_folder.clone());
+    let file = PathBuf::from(&hash[2..]);
+
+    fs::create_dir_all(&dir).context(format!("could not create all dir at {:?}", &dir))?;
+    let path = dir.join(file.clone());
+
+    if !std::path::Path::new(&path).exists() {
+        let mut f =
+            fs::File::create(&path).context(format!("could not create file at {:?}", &path))?;
+        f.write_all(contents.as_bytes())
+            .context("could not write bytes")?;
+    }
+
+    Ok(hash)
+
+
+/// Converts a hash string into a relative path like `c6/b8e869fa533155bbf2f0dd8fda9c68`.
+pub(crate) fn hash_to_path(hash: &str) -> Result<PathBuf> {
+    if hash.len() < 3 {
+        return Err(anyhow::anyhow!("Hash too short"));
+    }
+
+    let (first_two, rest) = hash.split_at(2);
+    Ok(PathBuf::from(first_two).join(rest))
+}
+
+/// Reads the file corresponding to the hash from the given base path.
+pub(crate) fn read_hash_file(base_path: &Path, hash: &str) -> Result<String> {
+    let relative_path = hash_to_path(hash)?;
+    let file_path = base_path.join(relative_path);
+    let contents = fs::read_to_string(file_path)?;
+
+    Ok(contents)
+}
+
+/// Walks through a folder, creating pinned entries as appropriate for every
+/// directory and file.  Returns a hash of the object.
+pub fn snapshot(store_path: &Path, dir: &Path) -> Result<String> {
+    if dir.is_dir() {
+        let mut entries: Vec<_> = fs::read_dir(dir)?.filter_map(Result::ok).collect();
+        entries.sort_by(|a, b| a.file_name().cmp(&b.file_name()));
+
+        let mut tree = Tree::default();
+
+        for entry in entries {
+            let path = entry.path();
+            let name = path
+                .file_name()
+                .unwrap_or_default()
+                .to_str()
+                .unwrap_or_default();
+            if path.is_dir() {
+                let branch = snapshot(store_path, &path)?;
+                tree.entries.push(Entry {
+                    kind: EntryKind::Tree,
+                    name: name.to_string(),
+                    hash: branch,
+                });
+            } else {
+                let hash = pin_file(store_path, &path)?;
+                tree.entries.push(Entry {
+                    kind: EntryKind::Blob,
+                    name: name.to_string(),
+                    hash,
+                });
+            }
+        }
+
+        let contents = toml::to_string(&tree).unwrap();
+        let hash = pin_contents(store_path, contents)?;
+
+        return Ok(hash);
+    }
+    Err(anyhow::anyhow!("store_path should be a folder"))
+}
